+++ lib/std/functional.lua
@@ -297,7 +297,7 @@ local M = {
   -- @return `true` if *x* can be called, otherwise `false`
   -- @usage
   -- if callable (functable) then functable (args) end
-  callable = X ("callable (any)", callable),
+  callable = X ("callable (?any)", callable),
 
   --- A rudimentary case statement.
   -- Match *with* against keys in *branches* table.