+++ NEWS.md
@@ -1,5 +1,8 @@
 # Stdlib NEWS - User visible changes
 
+## Noteworthy changes in release ?.? (????-??-??) [?]
+
+
 ## Noteworthy changes in release 41.1.1 (2015-01-31) [stable]
 
 ### Bug fixes