+++ NEWS.md
@@ -19,7 +19,8 @@
   - `std.getmetamethod` no longer rejects non-table subjects when
     `_DEBUG.argcheck` is set.
 
-  - `functional.bind` propagates nil valued arguments correctly.
+  - `functional.bind` and `functional.compose` propagate nil valued
+    arguments correctly.
 
 
 ## Noteworthy changes in release 41.1.1 (2015-01-31) [stable]